chore: Combine discard form with action form
