+++ crates/frontend/src/components.rs
@@ -15,7 +15,6 @@ pub mod dropdown;
 pub mod experiment;
 pub mod experiment_action_form;
 pub mod experiment_conclude_form;
-pub mod experiment_discard_form;
 pub mod experiment_form;
 pub mod experiment_group_form;
 pub mod experiment_ramp_form;
